requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0
chromadb>=0.4.0
openai>=1.0.0
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0
aiofiles>=23.0.0
jinja2>=3.1.0
